Code refactor for voters
