+++ services/core/shared/core/compat/sdk_v5/schema/votesAggregated.js
